refactor(ReservationStorage): Reservation을 받도록 수정
- 기본값인 0을 설정할 수 있도록 final 삭제 및 생성자 추가
