+++ src/main/java/roomescape/storage/ReservationStorage.java
@@ -2,11 +2,7 @@ package roomescape.storage;
 
 import org.springframework.stereotype.Component;
 import roomescape.domain.Reservation;
-import roomescape.dto.ReservationRequest;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicLong;
@@ -29,20 +25,10 @@ public class ReservationStorage {
         this.atomicLong = atomicLong;
     }
 
-    public Reservation save(ReservationRequest reservationRequest) {
-        Reservation reservation = fromRequest(reservationRequest);
+    public Reservation save(Reservation reservation) {
+        reservation.setId(atomicLong.incrementAndGet());
         reservations.add(reservation);
         return reservation;
-    }
-
-    private Reservation fromRequest(ReservationRequest reservationRequest) {
-        long id = atomicLong.incrementAndGet();
-
-        String name = reservationRequest.name();
-        LocalDate date = reservationRequest.date();
-        LocalTime time = reservationRequest.time();
-        LocalDateTime dateTime = LocalDateTime.of(date, time);
-        return new Reservation(id, name, dateTime);
     }
 
     public List<Reservation> findAllReservations() {
